+++ vercel-deployments/src/index.ts
@@ -1,4 +1,6 @@
 export { vercelDeploymentsPlugin } from './plugin.js'
-export type { Resolvable, VercelDeploymentsPluginConfig } from './types.js'
-export { resolveTarget } from './utilities/resolveTarget.js'
-export type { ResolvedTarget } from './utilities/resolveTarget.js'
+export type {
+  DeploymentTarget,
+  ResolveDeploymentTarget,
+  VercelDeploymentsPluginConfig,
+} from './types.js'